+++ src/card/cards/combat_constraints.rs
@@ -33,3 +33,4 @@ pub const URZA_S_TOWER: CardDefinitionId = CardDefinitionId(1832);
 pub const WORD_OF_BINDING: CardDefinitionId = CardDefinitionId(1833);
 pub const DEEP_SPAWN: CardDefinitionId = CardDefinitionId(1834);
 pub const MISHRA_S_WAR_MACHINE: CardDefinitionId = CardDefinitionId(1835);
+pub const PART_WATER: CardDefinitionId = CardDefinitionId(1836);